* Ajout de la commande 'Redémarrer' * Ajout de commentaires * Quelques corrections en général

+++ Serveur/Serveur_Synergie/Reseau/Paquets/paquetinsertiontexte.cpp
@@ -1,7 +1,7 @@
 #include "paquetinsertiontexte.h"
 
 PaquetInsertionTexte::PaquetInsertionTexte(Client* auteur, Fichier* fichier, QString texte, int position) :
-    BasePaquet((qint8)5)
+    BasePaquet(5)
 {
     *m_Stream << auteur->getID();
     *m_Stream << fichier->getID();
